[aot_inductor] Lightweight model runner
Pull Request resolved: https://github.com/pytorch/pytorch/pull/110158

It's useful to have a simple, lightweight way to run a model that adds
essentially no overhead to calling the model's generated `run_impl` method.
This C API is a super thin wrapper around AOTInductorModel: Create, Run, and
Delete are provided, and do very little work beyond dispatch to the appropriate
helpers.

Note the Create function also provides additional functionality beyond the
Container API; it allows the user to pass in a weight map defined in userland,
which is a requirement for several serving use cases.
ghstack-source-id: 202242949
@exported-using-ghexport

Differential Revision: [D49670711](https://our.internmc.facebook.com/intern/diff/D49670711/)